+++ docs/HISTORY.rst
@@ -8,6 +8,7 @@ Pending
 * Add ``__eq__`` to ``_Record``
 * Add ``is_het`` and ``is_variant`` to ``_Call``
 * Drop aggressive parse mode: we're always aggressive.
+* Add tabix fetch for single calls
 
 0.2.2 Release
 -------------